refactor abc347d

+++ ATC/abc347/d.cpp
@@ -1,5 +1,4 @@
-#include <bits/stdc++.h>
-
+#include <iostream>
 using namespace std;
 using ll=unsigned long long;
 
@@ -11,16 +10,10 @@ int main(void)
 	ll X=0,Y=0;
 	//Check 1 bit
 	for(ll i=0;i<61;i++){
-		if(C>>i & 1){
-			if(a>b){
-				X |= (ll)1<<i;
-				a--;
-			}else{
-				Y |= (ll)1<<i;
-				b--;
-			}
+		if((C>>i) & 1){
+			if(a>b){ X |= 1LL<<i; a--;}
+			else   { Y |= 1LL<<i; b--;}
 		}
-		//cout << "dbg:" << a << " " << b << " " << X << " " << Y << endl;
 	}
 
 	if(a!=b){
@@ -32,28 +25,16 @@ int main(void)
 	for(ll i=0;i<61;i++){
 		if((C>>i & 1) == 0){
 			if(a>0){
-				X |= (ll)1<<i;
-				Y |= (ll)1<<i;
+				X |= 1LL<<i;
+				Y |= 1LL<<i;
 				a--;
 			}
-
 			if(a==0) break;
 		}
-		//cout << "dbg:" << a << " " << X << " " << Y << endl;
 	}
 
 	if(a>0) cout << -1 << endl;
 	else cout << X << " " << Y << endl;
-
-	//test
-	//int cnt_a = 0, cnt_b = 0;
-	//for(int i=0;i<63;i++){
-	//	if (X>>i &1) cnt_a++;
-	//	if (Y>>i &1) cnt_b++;
-	//}
-	//cout << "a " << cnt_a<< endl;
-	//cout << "b " << cnt_b<< endl;
-	//cout << "X^Y==C " << (X^Y) << endl;
 
 	return 0;
 }
